requirements.txt
# Основные зависимости
python-telegram-bot>=20.0
redis[async]>=5.0.0
python-dotenv>=1.0.0
aiofiles>=23.0
Pillow>=10.0.0
openai>=1.0.0
pydantic>=2.0.0

# PostgreSQL
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Тестирование
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Линтинг и форматирование
black>=24.1.0
isort>=5.13.0
flake8>=7.0.0
mypy>=1.8.0

# Мониторинг
prometheus-client>=0.19.0

# Безопасность
cryptography>=42.0.0
python-jose>=3.3.0
passlib>=1.7.4

# Утилиты
python-dateutil>=2.8.2
pytz>=2024.1

aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
black==25.1.0
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.2.0
colorama==0.4.6
coverage==7.8.0
distlib==0.3.9
distro==1.9.0
fal_client==0.6.0
ffprobe==0.5
filelock==3.18.0
flake8==7.2.0
frozenlist==1.6.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
identify==2.6.10
idna==3.10
iniconfig==2.1.0
jiter==0.9.0
lxml==5.4.0
mccabe==0.7.0
multidict==6.4.3
mypy_extensions==1.1.0
nodeenv==1.9.1
numpy==2.2.5
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.5.0
pre_commit==4.2.0
propcache==0.3.1
pycodestyle==2.13.0
pydantic==2.11.3
pydantic_core==2.33.1
pydub==0.25.1
pyflakes==3.3.2
pyTelegramBotAPI==4.26.0
pytest==8.3.5
pytest-cov==6.1.1
python-docx==1.1.2
python-dotenv==1.1.0
python-telegram-bot==22.0
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
virtualenv==20.31.2
yarl==1.20.0

# Redis
redis[async]>=5.0.1

# Testing
pytest-asyncio>=0.26.0
pytest-cov>=6.1.1
